core/ocr/windows_ocr.py
"""
Windows OCR engine implementation using Windows.Media.Ocr.
"""

import logging
import ctypes
from PIL import Image
from .base import BaseOCREngine

logger = logging.getLogger(__name__)


class WindowsOCREngine(BaseOCREngine):
    """Windows OCR engine implementation."""

    def __init__(self):
        self._available = None

    def _check_availability(self) -> bool:
        """Check if Windows OCR API is available."""
        try:
            # Try to import winrt (Windows Runtime)
            import winrt.windows.media.ocr as ocr
            import winrt.windows.storage.streams as streams
            import winrt.windows.graphics.imaging as imaging
            return True
        except ImportError:
            # Fallback: try using ctypes to call Windows API directly
            try:
                # Check if we're on Windows
                import platform
                if platform.system() == "Windows":
                    return True
            except Exception:
                pass
            return False

    def recognize(self, image: Image.Image) -> str:
        """
        Recognize text in an image using Windows OCR API.

        Args:
            image: PIL Image to process

        Returns:
            Recognized text as a string
        """
        try:
            # Try winrt approach first
            try:
                import winrt.windows.media.ocr as ocr
                import winrt.windows.storage.streams as streams
                import winrt.windows.graphics.imaging as imaging
                import io

                # Convert PIL image to bytes
                img_bytes = io.BytesIO()
                image.save(img_bytes, format='PNG')
                img_bytes.seek(0)

                # This is a simplified implementation
                # In practice, you'd need to properly set up the Windows Runtime
                # to use the OCR API with the image data
                engine = ocr.OcrEngine.try_create_from_user_profile_languages()
                if engine:
                    # Note: Full implementation would require async handling
                    # and proper conversion of image data to Windows Runtime format
                    pass

            except ImportError:
                pass

            # Fallback: use pytesseract if available, otherwise return empty
            try:
                import pytesseract
                return pytesseract.image_to_string(image, lang='chi_sim+eng')
            except ImportError:
                logger.warning("Windows OCR not fully implemented, pytesseract not available")
                return ""

        except Exception as e:
            logger.error(f"Windows OCR recognition failed: {e}")
            return ""

    @property
    def name(self) -> str:
        return "Windows OCR"

    @property
    def is_available(self) -> bool:
        """Check if Windows OCR is available."""
        if self._available is None:
            self._available = self._check_availability()
        return self._available
