requirements.txt
PyQt6>=6.4.0
paddleocr>=2.6.0
paddlepaddle>=2.4.0
Pillow>=9.0.0
pyautogui>=0.9.53
pywin32>=304; sys_platform == "win32"
